improved style, refactored js
have grouped buttons more clearly and added a display bar, as well as changed colors of the app. also refactored code to move equalhelper to be called in operatorhelper, so that equal button can be of class operator. also added eventlisteners that change the buttons' style when they are held down